Moved TARGET_MODE to Credo specific memmap
