+++ sonic_platform_base/sonic_xcvr/mem_maps/credo/aec_800g.py
@@ -5,7 +5,16 @@
 """
 
 from ..public.cmisTargetFWUpgrade import CmisTargetFWUpgradeMemMap
+from ...fields.xcvr_field import (
+    NumberRegField,
+    RegGroupField,
+)
+from ...fields import consts
 
 class CmisAec800gMemMap(CmisTargetFWUpgradeMemMap):
-    # Vendor specific implementation to be added here
-    pass
+    def __init__(self, codes):
+        super().__init__(codes)
+
+        self.VENDOR_CUSTOM = RegGroupField(consts.VENDOR_CUSTOM,
+            NumberRegField(consts.TARGET_MODE, self.getaddr(0x0, 64), ro=False)
+        )
